refactor: 메인 entitiy 파일 common/entities로 경로이동
