+++ src/app.module.ts
@@ -6,10 +6,10 @@ import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { AuthModule } from './auth/auth.module';
 import { UsersModule } from './users/users.module';
-import { UsersEntity } from './users/entities/users.entity';
+import { UsersEntity } from './common/entities/users.entity';
 import { GoogleStrategy } from './auth/strategies/google.strategy';
 import { OrdersModule } from './orders/orders.module';
-import { OrdersEntity } from './orders/entities/_orders.entity';
+import { OrdersEntity } from './common/entities/orders.entity';
 import { ShopsModule } from './shops/shops.module';
 import { ProductsController } from './products/products.controller';
 import { ProductsModule } from './products/products.module';
