+++ src/common/entities/index.ts
@@ -1,1 +1,4 @@
+export { UsersEntity } from './users.entity';
+export { ShopsEntity } from './shops.entity';
+export { OrdersEntity } from './orders.entity';
 export { ReviewsEntity } from './reviews.entity';
