+++ src/common/entities/shops.entity.ts
@@ -1,6 +1,6 @@
 import { Column, Entity, ManyToOne, PrimaryGeneratedColumn } from 'typeorm';
 import { IsOptional, IsString, MaxLength } from 'class-validator';
-import { UsersEntity } from '../../users/entities/users.entity';
+import { UsersEntity } from './users.entity';
 
 @Entity({ name: 'shops' })
 export class ShopsEntity {
