+++ src/common/entities/users.entity.ts
@@ -1,8 +1,8 @@
 import { Column, Entity, OneToMany, PrimaryGeneratedColumn } from 'typeorm';
 import { Exclude } from 'class-transformer';
 import { IsOptional, IsString } from 'class-validator';
-import { OrdersEntity } from 'src/orders/entities/_orders.entity';
-import { ShopsEntity } from '../../shops/entities/shops.entity';
+import { OrdersEntity } from 'src/common/entities/orders.entity';
+import { ShopsEntity } from './shops.entity';
 
 export enum Provider {
   Local,
@@ -54,7 +54,7 @@ export class UsersEntity {
 
   @OneToMany(() => OrdersEntity, (order) => order.user)
   public orders: OrdersEntity[];
-  
+
   @OneToMany(() => ShopsEntity, (shop) => shop.user)
   public shops: ShopsEntity[];
 }
