+++ src/orders/dto/create-order.dto.ts
@@ -7,7 +7,7 @@ import {
   IsOptional,
   IsString,
 } from 'class-validator';
-import { OrderStatus } from '../entities/_orders.entity';
+import { OrderStatus } from '../../common/entities/orders.entity';
 
 export class OrderCreateDto {
   @IsNumber()
