+++ src/orders/orders.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { OrdersController } from './orders.controller';
 import { OrdersService } from './orders.service';
-import { OrdersEntity } from './entities/_orders.entity';
+import { OrdersEntity } from '../common/entities/orders.entity';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { UsersEntity } from 'src/users/entities/users.entity';
+import { UsersEntity } from 'src/common/entities/users.entity';
 import { UserService } from 'src/users/users.service';
 import { OrderDetailsEntity } from './entities/order-detail.entity';
