+++ src/reviews/reviews.module.ts
@@ -3,7 +3,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { ReviewsService } from './reviews.service';
 import { ReviewsController } from './reviews.controller';
 import { ReviewsEntity } from 'src/common/entities';
-import { UsersEntity } from 'src/users/entities/users.entity';
+import { UsersEntity } from 'src/common/entities/users.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([ReviewsEntity, UsersEntity])],
