+++ src/shops/shops.controller.ts
@@ -14,7 +14,7 @@ import { AuthGuard } from '../auth/auth.guard';
 import { ResultableInterface } from '../common/interfaces';
 import { RequestUserInterface } from '../users/interfaces';
 import { ShopCreateDto, ShopUpdateDto } from './dto/index';
-import { ShopsEntity } from './entities/shops.entity';
+import { ShopsEntity } from '../common/entities/shops.entity';
 import { ShopsService } from './shops.service';
 
 @Controller('shops')
