+++ src/shops/shops.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { ShopsService } from './shops.service';
 import { ShopsController } from './shops.controller';
-import { ShopsEntity } from './entities/shops.entity';
+import { ShopsEntity } from '../common/entities/shops.entity';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { UsersEntity } from '../users/entities/users.entity';
+import { UsersEntity } from '../common/entities/users.entity';
 import { UserService } from '../users/users.service';
 
 @Module({
