+++ src/users/users.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UserService } from './users.service';
 import { UserController } from './users.controller';
-import { UsersEntity } from './entities/users.entity';
+import { UsersEntity } from '../common/entities/users.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([UsersEntity])],
